requirements.txt
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
python-dotenv==0.19.2
whitenoise==6.0.0
Pillow==9.0.1
django-storages==1.12.3
boto3==1.21.31
django-extensions==3.1.5

pandas==1.4.2
openpyxl==3.0.9